refactor: split SemanticAnalysisVisitor into its own file
Separates the visitor (enums, MachineState, AST traversal logic) from
the SemanticAnalyzer service (orchestration + variable diagnostics)
for better single-responsibility adherence.
